Hidden the rup_loss_table output [skip hazardlib][demos]
